requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.2
praw==7.7.1
youtube-comment-downloader==0.1.78
pydantic-settings==2.4.0

playwright
TikTokApi
httpx
playwright
beautifulsoup4
